seperate game1 logic and display
